requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
bleach==6.0.0
email-validator==2.0.0
redis==4.6.0
celery==5.3.2